test: add stream delivery integ tests and TESTING.md
