+++ .github/workflows/integration-testing.yml
@@ -117,6 +117,12 @@ jobs:
             timeout: 10
             extra-deps: ""
             ignore: ""
+          # TODO: expand to full tests_integ/memory once test stability is addressed
+          - group: memory
+            path: tests_integ/memory/test_memory_client.py
+            timeout: 10
+            extra-deps: ""
+            ignore: ""
           - group: evaluation
             path: tests_integ/evaluation
             timeout: 15
@@ -151,6 +157,8 @@ jobs:
         env:
           AWS_REGION: us-west-2
           PYTHONUNBUFFERED: 1
+          MEMORY_KINESIS_ARN: ${{ secrets.MEMORY_KINESIS_ARN }}
+          MEMORY_ROLE_ARN: ${{ secrets.MEMORY_ROLE_ARN }}
         id: tests
         timeout-minutes: ${{ matrix.timeout }}
         run: |
